Make use of getTranslatedRemainder when translating multiple segments
Some translators' nextTranslatedRow may return a string that is
shorter than countRemaining even if countRemaining is less than
width. This may happen when it is desired to keep the last word of a
segment together with the next. Because a translator does not know
anything about the surrounding segments (see issue
https://github.com/brailleapps/dotify.api/issues/13), it must assume
that the first word of the next segment is too wide. So if we don't
take this into account we'll get premature line breaks.

By making use of getTranslatedRemainder we can find the first segment
that does not fit completely on the line, and look for break points
starting from that segment and going back towards the first.

Also fixed testHyphenateLastLine tests which were broken by the
previous commit for some reason.